Convert AR feature to TypeScript
