+++ src/types/document.d.ts
@@ -1,0 +1,4 @@
+interface Document {
+  // @see https://fullscreen.spec.whatwg.org/#dom-document-fullscreenelement
+  fullscreenElement: Element|null;
+}
